pyproject.toml
[tool.poetry]
name = "dog-bot"
version = "0.1.0"
description = ""
authors = ["Tucker Beck <tucker.beck@gmail.com>"]
readme = "README.md"
packages = [{include = "bot"}]

[tool.poetry.dependencies]
python = ">=3.11, <3.12"
discord-py = "^2.3.2"
loguru = "^0.7.1"
openai = "^0.28.0"
snick = "^1.3.0"
auto-name-enum = "^2.0.0"
pydantic-settings = "^2.0.3"
py-buzz = "^4.1.0"
python-statemachine = "^2.1.1"


[tool.poetry.scripts]
bot = "bot.main:run"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
